requirements.txt
django == 3.2.3
pillow == 8.2.0
django-admin-interface == 0.16.3
django-crispy-forms == 1.11.2
django-rest-framework
django-pwa 


## para importar la base de datos debes usar el comando python manage.py import